requirements/base.txt
tqdm==4.48.2
numpy==1.19.1
scipy==1.5.2
scikit-learn==0.23.2
pandas==1.1.0
pytest==6.0.1
pyhocon==0.3.55

torchvision==0.7.0
pytorch-lightning==0.9.0